refactor(controls): move controls to a dedicated components
